example01 using Time class

+++ cilantro/Time.cpp
@@ -1,7 +1,7 @@
 #include "Time.h"
 
-std::chrono::duration<double> Time::frameDeltaTime;
-std::chrono::duration<double> Time::frameRenderTime;
+std::chrono::duration<float> Time::frameDeltaTime;
+std::chrono::duration<float> Time::frameRenderTime;
 std::chrono::high_resolution_clock::time_point Time::timeOfFrameStart;
 std::chrono::high_resolution_clock::time_point Time::timeOfFrameEnd;
 
@@ -14,12 +14,12 @@ Time::~Time ()
 }
 
 
-double Time::GetFrameDeltaTime ()
+float Time::GetFrameDeltaTime ()
 {
 	return std::chrono::duration_cast<std::chrono::microseconds>(frameDeltaTime).count () / 1000000.0f;
 }
 
-double Time::GetFrameRenderTime ()
+float Time::GetFrameRenderTime ()
 {
 	return std::chrono::duration_cast<std::chrono::microseconds>(frameRenderTime).count () / 1000000.0f;
 }
